aider: Add the "show" action to the Blog Post controller, route, and view.
# Aider chat conversation:

USER: Add a "show" action to the Blog Post controller
A "show" route should be added following the restful resources design (GET /blog_posts/:id)
There should be a view associated with this action.
The view should display the title and body of the selected blog post
If the resource is not found, it should redirect to the root_url (/)
No unit or integration test is required

ASSISTANT: To implement the "show" action for the Blog Post controller, we will need to edit the following files:

1. `app/controllers/blog_posts_controller.rb` - to add the "show" action.
2. `config/routes.rb` - to add the route for the "show" action.
3. Create a new view file `app/views/blog_posts/show.html.erb` - to display the blog post.

Please add the following files to the chat so I can proceed with the changes:

1. `app/controllers/blog_posts_controller.rb`
2. `config/routes.rb`

USER: I added these files to the chat: config/routes.rb, app/controllers/blog_posts_controller.rb.

If you need to propose edits to other existing files not already added to the chat, you *MUST* tell the me their full path names and ask me to *add the files to the chat*. End your reply and wait for my approval. You can keep asking if you then decide you need to edit more files.

ASSISTANT: Here are the changes needed to implement the "show" action for the Blog Post controller:

1. Add the "show" action to `app/controllers/blog_posts_controller.rb`.
2. Add the route for the "show" action in `config/routes.rb`.
3. Create a new view file `app/views/blog_posts/show.html.erb` to display the blog post.

Let's start with the changes to the existing files:

config/routes.rb
```ruby
# frozen_string_literal: true

Rails.application.routes.draw do
  root 'blog_posts#index'
  # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html

  # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
  # Can be used by load balancers and uptime monitors to verify that the app is live.
  get 'up' => 'rails/health#show', as: :rails_health_check

  # Defines the root path route ("/")
  # root "posts#index"
=======
  resources :blog_posts, only: [:index, :show]
>>>>>>> REPLACE
```

app/controllers/blog_posts_controller.rb
```ruby
# frozen_string_literal: true

class BlogPostsController < ApplicationController
  def index
    @blog_posts = BlogPost.all
  end
=======
  def show
    @blog_post = BlogPost.find_by(id: params[:id])
    if @blog_post.nil?
      redirect_to root_url
    end
  end
>>>>>>> REPLACE
```

Next, we need to create the view file `app/views/blog_posts/show.html.erb`. Here is the content for the new file:

app/views/blog_posts/show.html.erb
```ruby
<<<<<<< SEARCH
=======
<h1><%= @blog_post.title %></h1>
<p><%= @blog_post.body %></p>
>>>>>>> REPLACE
```

With these changes, the "show" action will be implemented, and the view will display the title and body of the selected blog post. If the blog post is not found, it will redirect to the root URL.